+++ RELEASE_NOTES.rst
@@ -9,6 +9,7 @@ Features:
   :code:`chr1 1003 |GCGCGC|TTG`, where the bars are just visual aids to distinguish
   between the TRs and the flanking bps within the allele.) Without :code:`--trim`, such loci would
   be skipped due to partial overlaps.
+* qcSTR now has an option :code:`--outtype png` to produce pngs instead of pdfs
 
 Command line interface changes: